requirements.txt
asgiref==3.5.2
attrs==22.1.0
autopep8==1.6.0
black==22.6.0
Brotli==1.0.9
CacheControl==0.12.11
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
ConfigArgParse==1.5.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.4
distlib==0.3.5
Django
django-cors-headers==3.13.0
django-environ==0.9.0
django-paypal==2.0
djangorestframework==3.13.1
drf-stripe==1.2.6
drf-stripe-subscription==1.1.11
drf-yasg==1.21.0
exceptiongroup==1.0.0
filelock==3.7.1
firebase-admin==5.2.0
gevent==22.10.1
geventhttpclient==2.0.8
greenlet==1.1.3.post0
grpcio==1.47.0
grpcio-status==1.47.0
gunicorn==20.1.0
httplib2==0.20.4
idna==3.3
inflection==0.5.1
iniconfig==1.1.1
itsdangerous==2.1.2
itypes==1.2.0
Jinja2==3.1.2
jsonfield==3.1.0
jsonschema==4.9.1
locust==2.13.0
MarkupSafe==2.1.1
msgpack==1.0.4
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
platformdirs==2.5.2
pluggy==1.0.0
proto-plus==1.20.6
protobuf==3.20.1
psutil==5.9.3
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
pycryptodome==3.15.0
pydantic==1.9.2
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyrsistent==0.18.1
pytest==7.2.0
pytest-django==4.5.2
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
pyzmq==24.0.1
requests==2.28.1
roundrobin==0.0.4
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
stats==0.1.2a0
telebirr==0.0.4
toml==0.10.2
tomli==2.0.1
typing_extensions==4.3.0
uritemplate==4.1.1
urllib3==1.26.10
virtualenv==20.15.1
Werkzeug==2.2.2
wget==3.2
zope.event==4.5.0
zope.interface==5.5.0


azure-common==1.1.28
azure-core==1.26.1
azure-keyvault-secrets==4.6.0
azure-mgmt-core==1.3.2
azure-mgmt-storage==21.0.0
azure-storage-blob==2.1.0
azure-storage-common==2.1.0
# django-storages
django-storages-azure==1.6.8
# azure-storage-blob